Reuse Vec3 to optimize the Color structure

+++ src/utils/arithmetics.cpp
@@ -1,6 +1,6 @@
 #include "arithmetics.h"
 
-float Arithmetics::scale(float num,float in_min, float in_max,float out_min,float out_max)
+double Arithmetics::scale(double num,double in_min, double in_max,double out_min,double out_max)
 {
-    return (float) (num - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
+    return (double) (num - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
